text_analysis/requirements.txt
tensorflow-cpu>=2.17.0
gradio
joblib
